Added PubDevices model and query helper. Also updated DbUtils to better support foreign keys and multiple table joins.

+++ backend/src/test/java/com/sababado/mcpubs/backend/db/PubDevicesQueryHelperTest.java
@@ -3,6 +3,8 @@ package com.sababado.mcpubs.backend.db;
 import com.sababado.mcpubs.backend.db.utils.DbUtils;
 import com.sababado.mcpubs.backend.models.Device;
 import com.sababado.mcpubs.backend.models.Pub;
+import com.sababado.mcpubs.backend.models.PubDevices;
+import com.sababado.mcpubs.backend.utils.UnrecognizedPubException;
 
 import org.junit.After;
 import org.junit.Before;
@@ -12,6 +14,9 @@ import java.sql.Connection;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 /**
  * Created by robert on 9/1/16.
@@ -31,17 +36,27 @@ public class PubDevicesQueryHelperTest {
     }
 
     @Test
-    public void testInsertPubDevicesRecord() {
+    public void testInsertPubDevicesRecord() throws UnrecognizedPubException {
+        Pub pub = new Pub("MCO AAA4990.342B", "A readable title yeahyeah", true, Pub.MCO);
+        pub = PubQueryHelper.insertOrUpdateRecord(connection, pub);
+        String newToken = "AAA567980ghjklr7689";
+        Device device = DeviceQueryHelper.updateToken(connection, null, newToken);
 
+        PubDevices pubDevices = PubDevicesQueryHelper.insertPubDevicesRecord(connection, device.getId(), pub.getId());
+        assertTrue(pubDevices.getId() > 0);
+        pubDevicesIdsToCleanup.add(pubDevices.getId());
+        assertEquals(device, pubDevices.getDevice());
+        assertEquals(pub, pubDevices.getPub());
     }
 
     @After
     public void cleanup() {
         try {
-            String cleanupIds = Arrays.toString(pubDevicesIdsToCleanup.toArray(new Long[pubDevicesIdsToCleanup.size()]));
-            cleanupIds = cleanupIds.replace("[", "(").replace("]", ")");
-            connection.prepareStatement("DELETE FROM PubDevices WHERE ID IN " + cleanupIds + ";").execute();
-
+            if (pubDevicesIdsToCleanup.size() > 0) {
+                String cleanupIds = Arrays.toString(pubDevicesIdsToCleanup.toArray(new Long[pubDevicesIdsToCleanup.size()]));
+                cleanupIds = cleanupIds.replace("[", "(").replace("]", ")");
+                connection.prepareStatement("DELETE FROM PubDevices WHERE ID IN " + cleanupIds + ";").execute();
+            }
             connection.prepareStatement("DELETE FROM Device WHERE " + Device.DEVICE_TOKEN + " LIKE 'AAA%';").execute();
             connection.prepareStatement("DELETE FROM PUB WHERE " + Pub.FULL_CODE + " LIKE 'AAA%';").execute();
         } catch (Exception e) {
